Convert diminish and calls to use-package.
Using require would prevent Emacs from booting my
configuration. Moreover, using the :diminish marker within use-package
was not diminishing reliably. Not sure if this is on my end,
the end of the packages, or of use-package. At any rate, using
use-package to call diminish (at the end of the configuration, as
recommended by Will Mengarini) and then making specific commands on
:init works for now. This may prove unsustainable, however, if I want to
diminish minor modes that are not always present (unlike smartparens and
company-mode).